Fix tests execution
